requirements.txt
# Flask web framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1

# PDF generation
reportlab==4.0.8
pypdf==3.17.4

# Excel export
openpyxl==3.1.2

# Form validation
WTForms==3.1.1
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0
